@@ -11,20 +11,18 @@
     "@babel/runtime": "^7.22.6",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@fontsource/roboto": "^5.0.5",
+    "@mui/icons-material": "^5.14.0",
+    "@mui/material": "^5.14.0",
+    "@mui/x-data-grid": "^6.10.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
     "@testing-library/user-event": "^12.8.3",
     "file-saver": "^2.0.5",
-    "husky": "^7.0.4",
     "jszip": "^3.10.1",
-    "konva": "^7.2.5",
-    "lint-staged": "^10.5.4",
-    "material-ui-popup-state": "^2.0.1",
-    "notistack": "^1.0.10",
+    "konva": "^8.4.3",
+    "material-ui-popup-state": "^5.0.9",
+    "notistack": "^3.0.1",
     "qs": "^6.11.2",
     "react": "^17.0.2",
     "react-debounce-input": "^3.3.0",
@@ -33,10 +31,10 @@
     "react-joyride": "^2.5.4",
     "react-konva": "^17.0.2-6",
     "react-material-ui-form-validator": "^3.0.1",
-    "react-router-dom": "^6.14.1",
+    "react-router-dom": "^6.14.2",
     "socket.io-client": "^4.7.1",
-    "typescript": "^5.1.0",
-    "use-image": "^1.1.1"
+    "use-image": "^1.1.1",
+    "uuid": "^9.0.0"
   },
   "scripts": {
     "start": "node generate-env-browser.js && cp env-config.js ./public/ && vite",
@@ -69,29 +67,33 @@
     "@playwright/test": "^1.36.2",
     "@types/file-saver": "^2.0.5",
     "@types/jest": "^29.5.2",
-    "@types/lodash": "^4.14.195",
-    "@types/node": "^17.0.45",
+    "@types/node": "^18.16.19",
     "@types/qs": "^6.9.7",
     "@types/react": "^17.0.62",
     "@types/react-dom": "^17.0.20",
     "@types/react-material-ui-form-validator": "^2.1.1",
     "@types/react-router-dom": "^5.3.3",
-    "@types/testing-library__jest-dom": "^5.14.7",
-    "@typescript-eslint/eslint-plugin": "^5.61.0",
-    "@typescript-eslint/parser": "^5.61.0",
+    "@types/testing-library__jest-dom": "^5.14.8",
+    "@types/uuid": "^9.0.2",
+    "@typescript-eslint/eslint-plugin": "^6.1.0",
+    "@typescript-eslint/parser": "^6.1.0",
     "@visual-regression-tracker/agent-playwright": "^5.3.0",
-    "@vitejs/plugin-react": "^4.0.2",
-    "eslint": "^8.44.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-typescript": "^3.5.5",
     "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "husky": "^8.0.3",
     "jest": "^29.6.1",
     "jest-environment-jsdom": "^29.6.1",
+    "lint-staged": "^13.2.3",
     "prettier": "^3.0.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
+    "typescript": "^5.1.6",
     "vite": "^4.4.2"
   }
 }
