@@ -1,4 +1,4 @@
-import { FormControlLabel, Switch } from "@material-ui/core";
+import { FormControlLabel, Switch } from "@mui/material";
 import React from "react";
 import { TextValidator } from "react-material-ui-form-validator";
 import { OdiffConfig } from "../../types/imageComparison";
@@ -50,7 +50,13 @@ export const OdiffConfigForm: React.FunctionComponent = () => {
         name="threshold"
         validators={["minNumber:0", "maxNumber:1"]}
         errorMessages={["Enter greater than 0", "Enter less than 1"]}
-        InputProps={{ inputProps: { min: 0, max: 1, step: 0.001 } }}
+        InputProps={{
+          inputProps: {
+            min: 0,
+            max: 1,
+            step: 0.001,
+          },
+        }}
         margin="dense"
         id="threshold"
         label="Pixel diff threshold"
@@ -58,8 +64,8 @@ export const OdiffConfigForm: React.FunctionComponent = () => {
         fullWidth
         required
         value={config.threshold}
-        onChange={(event) => {
-          const value = (event.target as HTMLInputElement).value;
+        onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
+          const { value } = event.target;
           updateConfig("threshold", parseFloat(value));
         }}
       />
