@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import { styled } from "@mui/material/styles";
 import {
   Typography,
   Button,
@@ -6,15 +7,15 @@ import {
   Switch,
   IconButton,
   Box,
-  makeStyles,
   Select,
   MenuItem,
   LinearProgress,
   Divider,
   FormControlLabel,
   Checkbox,
-} from "@material-ui/core";
-import { ToggleButton } from "@material-ui/lab";
+  ToggleButton,
+  SelectChangeEvent,
+} from "@mui/material";
 import { useHotkeys } from "react-hotkeys-hook";
 import { TestRun } from "../../types";
 import { testRunService, staticService } from "../../services";
@@ -36,72 +37,97 @@ import {
   FullscreenExit,
   NavigateNext,
   NavigateBefore,
-} from "@material-ui/icons";
+} from "@mui/icons-material";
 import { TestRunDetails } from "./TestRunDetails";
 import useImage from "use-image";
 import { routes } from "../../constants";
 import { useTestRunDispatch } from "../../contexts";
-import { DrawArea } from "./DrawArea";
+import { DrawArea, ImageStateLoad } from "./DrawArea";
 import { CommentsPopper } from "../CommentsPopper";
 import { useSnackbar } from "notistack";
 import { ApproveRejectButtons } from "./ApproveRejectButtons";
-import { head } from "lodash";
 import { invertIgnoreArea } from "../../_helpers/ignoreArea.helper";
 import { BaseModal } from "../BaseModal";
 import { Tooltip } from "../Tooltip";
-import ImageDetails from "./ImageDetails";
+import ImageDetails, { ImageDetailsProps } from "./ImageDetails";
 import { calculateScale } from "../../_helpers/scale.helper";
 import TestStatusChip from "../TestStatusChip";
+
+const PREFIX = "TestDetailsModal";
+
+const classes = {
+  header: `${PREFIX}-header`,
+  footer: `${PREFIX}-footer`,
+  scaleActions: `${PREFIX}-scaleActions`,
+  testRunActions: `${PREFIX}-testRunActions`,
+  testRunName: `${PREFIX}-testRunName`,
+  closeIcon: `${PREFIX}-closeIcon`,
+  testRunDetails: `${PREFIX}-testRunDetails`,
+  drawAreaContainer: `${PREFIX}-drawAreaContainer`,
+  drawAreaItem: `${PREFIX}-drawAreaItem`,
+  imageToolbar: `${PREFIX}-imageToolbar`,
+};
+
+// TODO jss-to-styled codemod: The Fragment root was replaced by div. Change the tag if needed.
+const Root = styled("div")(() => ({
+  [`& .${classes.header}`]: {
+    position: "relative",
+    textAlign: "left",
+    background: "#efefef",
+    paddingLeft: 8,
+    paddingBottom: 8,
+  },
+
+  [`& .${classes.footer}`]: {
+    background: "#efefef",
+  },
+
+  [`& .${classes.scaleActions}`]: {
+    display: "flex",
+    alignItems: "center",
+  },
+
+  [`& .${classes.testRunActions}`]: {
+    display: "flex",
+    alignItems: "center",
+    alignContent: "center",
+  },
+
+  [`& .${classes.testRunName}`]: {
+    fontWeight: 300,
+  },
+
+  [`& .${classes.closeIcon}`]: {
+    position: "absolute",
+    right: "8px",
+  },
+
+  [`& .${classes.testRunDetails}`]: {
+    paddingLeft: 8,
+  },
+
+  [`& .${classes.drawAreaContainer}`]: {
+    width: "100%",
+    height: "100%",
+  },
+
+  [`& .${classes.drawAreaItem}`]: {
+    padding: "0 4px",
+    height: "100%",
+  },
+
+  [`& .${classes.imageToolbar}`]: {
+    paddingLeft: 5,
+    height: 52,
+  },
+}));
 
 const defaultStagePos = {
   x: 0,
   y: 0,
 };
 
-const useStyles = makeStyles((theme) => ({
-  header: {
-    position: "relative",
-    textAlign: "left",
-    background: "#efefef",
-    paddingLeft: 8,
-  },
-  footer: {
-    background: "#efefef",
-  },
-  scaleActions: {
-    display: "flex",
-    alignItems: "center",
-  },
-  testRunActions: {
-    display: "flex",
-    alignItems: "center",
-    alignContent: "center",
-  },
-  testRunName: {
-    fontWeight: 300,
-  },
-  closeIcon: {
-    position: "absolute",
-    right: "8px",
-  },
-  testRunDetails: {
-    paddingLeft: 8,
-  },
-  drawAreaContainer: {
-    width: "100%",
-    height: "100%",
-  },
-  drawAreaItem: {
-    padding: "0 4px",
-    height: "100%",
-  },
-  imageToolbar: {
-    paddingLeft: 5,
-    height: 52,
-  },
-}));
-
-const TestDetailsModal: React.FunctionComponent<{
+interface TestDetailsModalProps {
   testRun: TestRun;
   currentRunIndex: number;
   totalTestRunCount: number;
@@ -109,7 +135,9 @@ const TestDetailsModal: React.FunctionComponent<{
   handleClose: () => void;
   handlePrevious: () => void;
   handleNext: () => void;
-}> = ({
+}
+
+const TestDetailsModal: React.FunctionComponent<TestDetailsModalProps> = ({
   testRun,
   currentRunIndex,
   totalTestRunCount,
@@ -118,7 +146,6 @@ const TestDetailsModal: React.FunctionComponent<{
   handleNext,
   handleClose,
 }) => {
-  const classes = useStyles();
   const { enqueueSnackbar } = useSnackbar();
   const testRunDispatch = useTestRunDispatch();
 
@@ -130,21 +157,27 @@ const TestDetailsModal: React.FunctionComponent<{
   const [stageOffset, setStageOffset] = React.useState(defaultStagePos);
   const [processing, setProcessing] = React.useState(false);
   const [isDrawMode, setIsDrawMode] = useState(false);
+
   const [valueOfIgnoreOrCompare, setValueOfIgnoreOrCompare] =
     useState("Ignore Areas");
+
   const [isDiffShown, setIsDiffShown] = useState(false);
   const [selectedRectId, setSelectedRectId] = React.useState<string>();
   const [ignoreAreas, setIgnoreAreas] = React.useState<IgnoreArea[]>([]);
+
   const [applyIgnoreDialogOpen, setApplyIgnoreDialogOpen] =
     React.useState(false);
 
   const GO_TO_NEXT_KEY = "goToNextAutomatically";
+
   const [goToNextAutomatically, setGoToNextAutomatically] = React.useState(
     () => {
       const json = localStorage.getItem(GO_TO_NEXT_KEY);
+
       if (json) {
         return JSON.parse(json);
       }
+
       return false;
     },
   );
@@ -166,13 +199,14 @@ const TestDetailsModal: React.FunctionComponent<{
     setApplyIgnoreDialogOpen(!applyIgnoreDialogOpen);
   };
 
-  console.log("rerender TestDetailsModal");
-  const [image, imageStatus] = useImage(
+  const [currentImage, currentImageStatus] = useImage(
     staticService.getImage(testRun.imageName),
   );
+
   const [baselineImage, baselineImageStatus] = useImage(
     staticService.getImage(testRun.baselineName),
   );
+
   const [diffImage, diffImageStatus] = useImage(
     staticService.getImage(testRun.diffName),
   );
@@ -187,33 +221,36 @@ const TestDetailsModal: React.FunctionComponent<{
       container.offsetWidth,
       container.offsetHeight - 48,
     );
+
     if (scale < stageScale) {
       setStageScale(scale);
     }
   };
 
   const fitImagesToStage = () => {
-    if (image && leftItemRef.current) {
-      fitImageToStage(image, leftItemRef.current);
+    if (currentImage && leftItemRef.current) {
+      fitImageToStage(currentImage, leftItemRef.current);
     }
+
     if (baselineImage && rightItemRef.current) {
       fitImageToStage(baselineImage, rightItemRef.current);
     }
+
     resetPosition();
   };
 
-  React.useEffect(() => {
+  useEffect(() => {
     setIsDiffShown(!!testRun.diffName);
   }, [testRun.diffName]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     setIgnoreAreas(JSON.parse(testRun.ignoreAreas));
   }, [testRun]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     fitImagesToStage();
-    // eslint-disable-next-line
-  }, [baselineImage, image]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [baselineImage, currentImage]);
 
   useHotkeys("right", handleNext, [currentRunIndex, handleNext]);
   useHotkeys("left", handlePrevious, [currentRunIndex, handlePrevious]);
@@ -222,9 +259,9 @@ const TestDetailsModal: React.FunctionComponent<{
     () =>
       testRun.baselineName &&
       testRun.imageName &&
-      (image?.height !== baselineImage?.height ||
-        image?.width !== baselineImage?.width),
-    [image, baselineImage, testRun.baselineName, testRun.imageName],
+      (currentImage?.height !== baselineImage?.height ||
+        currentImage?.width !== baselineImage?.width),
+    [currentImage, baselineImage, testRun.baselineName, testRun.imageName],
   );
 
   const handleIgnoreAreaChange = (ignoreAreas: IgnoreArea[]) => {
@@ -238,6 +275,7 @@ const TestDetailsModal: React.FunctionComponent<{
   const removeSelection = (event: Konva.KonvaEventObject<MouseEvent>) => {
     // deselect when clicked not on Rect
     const isRectClicked = event.target.className === "Rect";
+
     if (!isRectClicked) {
       setSelectedRectId(undefined);
     }
@@ -269,11 +307,11 @@ const TestDetailsModal: React.FunctionComponent<{
   const saveIgnoreAreasOrCompareArea = () => {
     if (valueOfIgnoreOrCompare.includes("Ignore")) {
       saveTestRun(ignoreAreas, "Ignore areas are updated.");
-    } else {
+    } else if (currentImage) {
       const invertedIgnoreAreas = invertIgnoreArea(
-        image!.width,
-        image!.height,
-        head(ignoreAreas),
+        currentImage.width,
+        currentImage.height,
+        ignoreAreas[0],
       );
 
       handleIgnoreAreaChange(invertedIgnoreAreas);
@@ -282,7 +320,11 @@ const TestDetailsModal: React.FunctionComponent<{
         "Selected area has been inverted to ignore areas and saved.",
       );
     }
-    testRunDispatch({ type: "touched", payload: false });
+
+    testRunDispatch({
+      type: "touched",
+      payload: false,
+    });
   };
 
   const onIgnoreOrCompareSelectChange = (value: string) => {
@@ -305,8 +347,9 @@ const TestDetailsModal: React.FunctionComponent<{
 
   const applyIgnoreArea = () => {
     const newIgnoreArea = ignoreAreas.find(
-      (area) => selectedRectId! === area.id,
+      (area) => selectedRectId === area.id,
     );
+
     if (newIgnoreArea) {
       setProcessing(true);
       testRunService
@@ -315,8 +358,9 @@ const TestDetailsModal: React.FunctionComponent<{
           const allIds = testRuns.map((item) => item.id);
           const data: UpdateIgnoreAreaDto = {
             ids: allIds,
-            ignoreAreas: [newIgnoreArea!],
+            ignoreAreas: [newIgnoreArea],
           };
+
           testRunService.addIgnoreAreas(data).then(() => {
             setProcessing(false);
             setSelectedRectId(undefined);
@@ -329,7 +373,7 @@ const TestDetailsModal: React.FunctionComponent<{
           });
         })
         .catch((error) => {
-          enqueueSnackbar("There was an error : " + error, {
+          enqueueSnackbar(`There was an error : ${error}`, {
             variant: "error",
           });
           setProcessing(false);
@@ -337,14 +381,20 @@ const TestDetailsModal: React.FunctionComponent<{
     } else {
       enqueueSnackbar(
         "There was an error determining which ignore area to apply.",
-        { variant: "error" },
+        {
+          variant: "error",
+        },
       );
     }
   };
 
   useHotkeys(
     "d",
-    () => !!testRun.diffName && setIsDiffShown((isDiffShown) => !isDiffShown),
+    () => {
+      if (testRun.diffName) {
+        setIsDiffShown((isDiffShown) => !isDiffShown);
+      }
+    },
     [testRun.diffName],
   );
   useHotkeys("ESC", handleClose, [handleClose]);
@@ -356,186 +406,130 @@ const TestDetailsModal: React.FunctionComponent<{
     );
   };
 
-  const ignoreAreasToolbar = () => {
-    return (
-      <>
-        <Grid container alignItems="center" spacing={2}>
+  const ignoreAreasToolbar = () => (
+    <React.Fragment>
+      <Grid container alignItems="center" spacing={2}>
+        <Grid item>
+          <Select
+            variant="standard"
+            id="area-select"
+            labelId="areaSelect"
+            value={valueOfIgnoreOrCompare}
+            onChange={(event: SelectChangeEvent<HTMLInputElement>) =>
+              onIgnoreOrCompareSelectChange(event.target.value as string)
+            }
+          >
+            {["Ignore Areas", "Compare Area"].map((eachItem) => (
+              <MenuItem key={eachItem} value={eachItem}>
+                {eachItem}
+              </MenuItem>
+            ))}
+          </Select>
+        </Grid>
+        <Grid item>
+          <ToggleButton
+            value={"drawMode"}
+            selected={isDrawMode}
+            size="small"
+            onClick={() => {
+              setIsDrawMode(!isDrawMode);
+            }}
+            style={{
+              padding: 4,
+            }}
+          >
+            <Add />
+          </ToggleButton>
+        </Grid>
+        <Grid item>
+          <IconButton
+            size="small"
+            disabled={!selectedRectId || ignoreAreas.length === 0}
+            onClick={() => selectedRectId && deleteIgnoreArea(selectedRectId)}
+          >
+            <Delete />
+          </IconButton>
+        </Grid>
+        <Tooltip title="Clears all ignore areas." aria-label="reject">
           <Grid item>
-            <Select
-              id="area-select"
-              labelId="areaSelect"
-              value={valueOfIgnoreOrCompare}
-              onChange={(event) =>
-                onIgnoreOrCompareSelectChange(event.target.value as string)
-              }
-            >
-              {["Ignore Areas", "Compare Area"].map((eachItem) => (
-                <MenuItem key={eachItem} value={eachItem}>
-                  {eachItem}
-                </MenuItem>
-              ))}
-            </Select>
-          </Grid>
-          <Grid item>
-            <ToggleButton
-              value={"drawMode"}
-              selected={isDrawMode}
+            <IconButton
               size="small"
+              disabled={ignoreAreas.length === 0}
               onClick={() => {
-                setIsDrawMode(!isDrawMode);
+                handleIgnoreAreaChange([]);
               }}
-              style={{ padding: 4 }}
             >
-              <Add />
-            </ToggleButton>
+              <LayersClear />
+            </IconButton>
           </Grid>
+        </Tooltip>
+        <Tooltip title={applyIgnoreAreaText} aria-label="apply ignore area">
           <Grid item>
             <IconButton
               size="small"
               disabled={!selectedRectId || ignoreAreas.length === 0}
-              onClick={() => selectedRectId && deleteIgnoreArea(selectedRectId)}
+              onClick={() => toggleApplyIgnoreDialogOpen()}
             >
-              <Delete />
+              <Collections />
             </IconButton>
           </Grid>
-          <Tooltip title="Clears all ignore areas." aria-label="reject">
-            <Grid item>
-              <IconButton
-                size="small"
-                disabled={ignoreAreas.length === 0}
-                onClick={() => {
-                  handleIgnoreAreaChange([]);
-                }}
-              >
-                <LayersClear />
-              </IconButton>
-            </Grid>
-          </Tooltip>
-          <Tooltip title={applyIgnoreAreaText} aria-label="apply ignore area">
-            <Grid item>
-              <IconButton
-                size="small"
-                disabled={!selectedRectId || ignoreAreas.length === 0}
-                onClick={() => toggleApplyIgnoreDialogOpen()}
-              >
-                <Collections />
-              </IconButton>
-            </Grid>
-          </Tooltip>
-          <Grid item>
-            <IconButton
-              size="small"
-              disabled={!touched}
-              onClick={() => saveIgnoreAreasOrCompareArea()}
-            >
-              <Save />
-            </IconButton>
-          </Grid>
-        </Grid>
-      </>
-    );
-  };
-
-  const baselinePanel = () => {
-    return (
-      <Grid
-        container
-        xs={6}
-        ref={leftItemRef}
-        className={classes.drawAreaItem}
-        direction="column"
-        wrap="nowrap"
-        alignItems="stretch"
-      >
-        <Grid
-          container
-          alignItems="center"
-          spacing={2}
-          className={classes.imageToolbar}
-        >
-          <ImageDetails
-            type="Baseline"
-            branchName={testRun.baselineBranchName}
-            imageName={testRun.baselineName}
-            image={baselineImage}
-            ignoreAreas={[]}
-          />
-          <Grid item>
-            <Button
-              color="primary"
-              disabled={!testRun.testVariationId}
-              onClick={openHistoryTab}
-            >
-              History <OpenInNew fontSize="small" />
-            </Button>
-          </Grid>
-        </Grid>
-        <Grid item style={{ flexGrow: "1" }}>
-          <DrawArea
-            imageName={testRun.baselineName}
-            imageState={[baselineImage, baselineImageStatus]}
-            ignoreAreas={[]}
-            tempIgnoreAreas={[]}
-            setIgnoreAreas={handleIgnoreAreaChange}
-            selectedRectId={selectedRectId}
-            setSelectedRectId={setSelectedRectId}
-            onStageClick={removeSelection}
-            stageScaleState={[stageScale, setStageScale]}
-            stagePosState={[stagePos, setStagePos]}
-            stageScrollPosState={[stageScrollPos, setStageScrollPos]}
-            stageInitPosState={[stageInitPos, setStageInitPos]}
-            stageOffsetState={[stageOffset, setStageOffset]}
-            drawModeState={[false, setIsDrawMode]}
-          />
+        </Tooltip>
+        <Grid item>
+          <IconButton
+            size="small"
+            disabled={!touched}
+            onClick={() => saveIgnoreAreasOrCompareArea()}
+          >
+            <Save />
+          </IconButton>
         </Grid>
       </Grid>
-    );
-  };
+    </React.Fragment>
+  );
 
-  const diffPanel = (
-    type: any,
-    branchName: string,
-    imageName: string,
-    imageStatus: any,
-    image: undefined | HTMLImageElement,
-  ) => {
-    return (
-      <Grid item xs={6} className={classes.drawAreaItem} ref={rightItemRef}>
-        <Grid
-          item
-          container
-          alignItems="center"
-          spacing={2}
-          className={classes.imageToolbar}
-        >
-          <ImageDetails
-            type={type}
-            branchName={branchName}
-            imageName={imageName}
-            image={image}
-            ignoreAreas={JSON.parse(testRun.tempIgnoreAreas)}
-          />
-          {testRun.diffName && (
-            <Grid item style={{ padding: 0 }}>
-              <Tooltip title={"Toggle diff. Hotkey: D"}>
-                <Switch
-                  checked={isDiffShown}
-                  onChange={() => setIsDiffShown(!isDiffShown)}
-                  name="Toggle diff"
-                />
-              </Tooltip>
-            </Grid>
-          )}
-          <Grid item>{ignoreAreasToolbar()}</Grid>
+  const baselinePanel = () => (
+    <Grid
+      item
+      ref={leftItemRef}
+      className={classes.drawAreaItem}
+      alignItems="stretch"
+    >
+      <Grid
+        container
+        alignItems="center"
+        spacing={2}
+        className={classes.imageToolbar}
+      >
+        <ImageDetails
+          type="Baseline"
+          branchName={testRun.baselineBranchName}
+          imageName={testRun.baselineName}
+          image={baselineImage}
+          ignoreAreas={[]}
+        />
+        <Grid item xs={6}>
+          <Button
+            color="primary"
+            disabled={!testRun.testVariationId}
+            onClick={openHistoryTab}
+          >
+            History <OpenInNew fontSize="small" />
+          </Button>
         </Grid>
+      </Grid>
+      <Grid
+        item
+        style={{
+          flexGrow: "1",
+        }}
+      >
         <DrawArea
-          imageName={imageName}
-          imageState={[image, imageStatus]}
-          ignoreAreas={ignoreAreas}
-          tempIgnoreAreas={JSON.parse(testRun.tempIgnoreAreas)}
+          imageName={testRun.baselineName}
+          imageState={[baselineImage, baselineImageStatus]}
+          ignoreAreas={[]}
+          tempIgnoreAreas={[]}
           setIgnoreAreas={handleIgnoreAreaChange}
           selectedRectId={selectedRectId}
-          deleteIgnoreArea={deleteIgnoreArea}
           setSelectedRectId={setSelectedRectId}
           onStageClick={removeSelection}
           stageScaleState={[stageScale, setStageScale]}
@@ -543,98 +537,153 @@ const TestDetailsModal: React.FunctionComponent<{
           stageScrollPosState={[stageScrollPos, setStageScrollPos]}
           stageInitPosState={[stageInitPos, setStageInitPos]}
           stageOffsetState={[stageOffset, setStageOffset]}
-          drawModeState={[isDrawMode, setIsDrawMode]}
+          drawModeState={[false, setIsDrawMode]}
         />
       </Grid>
-    );
-  };
+    </Grid>
+  );
 
-  const header = () => {
-    return (
-      <Box m={1}>
-        <Grid
-          container
-          alignItems="center"
-          className={classes.header}
-          spacing={2}
-        >
-          <Grid item>
-            <Typography variant="h6" display="inline">{`Step ${
-              currentRunIndex + 1
-            }/${totalTestRunCount}: `}</Typography>
-            <Typography
-              variant="h6"
-              display="inline"
-              className={classes.testRunName}
-            >
-              {testRun.name}
-            </Typography>
-          </Grid>
-          <Grid item>
-            <TestStatusChip status={testRun.status} />
-          </Grid>
-          <IconButton
-            color="inherit"
-            onClick={handleClose}
-            className={classes.closeIcon}
+  const diffPanel = (
+    type: ImageDetailsProps["type"],
+    branchName: string,
+    imageName: string,
+    imageStatus: ImageStateLoad,
+    image: HTMLImageElement | undefined,
+  ) => (
+    <Grid item xs={6} className={classes.drawAreaItem} ref={rightItemRef}>
+      <Grid
+        container
+        alignItems="center"
+        spacing={2}
+        className={classes.imageToolbar}
+      >
+        <ImageDetails
+          type={type}
+          branchName={branchName}
+          imageName={imageName}
+          image={image}
+          ignoreAreas={JSON.parse(testRun.tempIgnoreAreas)}
+        />
+        {testRun.diffName && (
+          <Grid
+            item
+            style={{
+              padding: 0,
+            }}
           >
+            <Tooltip title={"Toggle diff. Hotkey: D"}>
+              <Switch
+                checked={isDiffShown}
+                onChange={() => setIsDiffShown(!isDiffShown)}
+                name="Toggle diff"
+              />
+            </Tooltip>
+          </Grid>
+        )}
+        <Grid item>{ignoreAreasToolbar()}</Grid>
+      </Grid>
+      <DrawArea
+        imageName={imageName}
+        imageState={[image, imageStatus]}
+        ignoreAreas={ignoreAreas}
+        tempIgnoreAreas={JSON.parse(testRun.tempIgnoreAreas)}
+        setIgnoreAreas={handleIgnoreAreaChange}
+        selectedRectId={selectedRectId}
+        deleteIgnoreArea={deleteIgnoreArea}
+        setSelectedRectId={setSelectedRectId}
+        onStageClick={removeSelection}
+        stageScaleState={[stageScale, setStageScale]}
+        stagePosState={[stagePos, setStagePos]}
+        stageScrollPosState={[stageScrollPos, setStageScrollPos]}
+        stageInitPosState={[stageInitPos, setStageInitPos]}
+        stageOffsetState={[stageOffset, setStageOffset]}
+        drawModeState={[isDrawMode, setIsDrawMode]}
+      />
+    </Grid>
+  );
+
+  const header = () => (
+    <Box mt={1} ml={1}>
+      <Grid
+        container
+        alignItems="center"
+        className={classes.header}
+        spacing={2}
+      >
+        <Grid item>
+          <Typography variant="h6" display="inline">{`Step ${
+            currentRunIndex + 1
+          }/${totalTestRunCount}: `}</Typography>
+          <Typography
+            variant="h6"
+            display="inline"
+            className={classes.testRunName}
+          >
+            {testRun.name}
+          </Typography>
+        </Grid>
+        <Grid item>
+          <TestStatusChip status={testRun.status} />
+        </Grid>
+        <Grid item className={classes.closeIcon}>
+          <IconButton color="inherit" onClick={handleClose} size="large">
             <Close />
           </IconButton>
         </Grid>
-      </Box>
-    );
-  };
+      </Grid>
+    </Box>
+  );
 
-  const testRunDetails = () => {
-    return (
-      <Box ml={1} mr={1} mt={0} mb={0}>
-        <Grid
-          container
-          alignItems="center"
-          className={classes.testRunDetails}
-          spacing={1}
-        >
-          <TestRunDetails testRun={testRun} />
-          {isImageSizeDiffer && (
-            <Grid item>
-              <Tooltip
-                title={
-                  "Image height/width differ from baseline! Cannot calculate diff!"
-                }
-              >
-                <IconButton>
-                  <WarningRounded color="secondary" />
-                </IconButton>
-              </Tooltip>
-            </Grid>
-          )}
+  const testRunDetails = () => (
+    <Box ml={1} mr={1} mt={0} mb={0}>
+      <Grid
+        container
+        alignItems="center"
+        className={classes.testRunDetails}
+        spacing={1}
+      >
+        <TestRunDetails testRun={testRun} />
+        {isImageSizeDiffer && (
           <Grid item>
-            <CommentsPopper
-              text={testRun.comment}
-              onSave={(comment) =>
-                testRunService
-                  .update(testRun.id, { comment })
-                  .then(() =>
-                    enqueueSnackbar("Comment updated", {
-                      variant: "success",
-                    }),
-                  )
-                  .catch((err) =>
-                    enqueueSnackbar(err, {
-                      variant: "error",
-                    }),
-                  )
+            <Tooltip
+              title={
+                "Image height/width differ from baseline! Cannot calculate diff!"
               }
-            />
+            >
+              <IconButton size="large">
+                <WarningRounded color="secondary" />
+              </IconButton>
+            </Tooltip>
           </Grid>
+        )}
+        <Grid item>
+          <CommentsPopper
+            text={testRun.comment}
+            onSave={(comment) =>
+              testRunService
+                .update(testRun.id, {
+                  comment,
+                })
+                .then(() =>
+                  enqueueSnackbar("Comment updated", {
+                    variant: "success",
+                  }),
+                )
+                .catch((err) =>
+                  enqueueSnackbar(err, {
+                    variant: "error",
+                  }),
+                )
+            }
+          />
         </Grid>
-        <Divider />
-      </Box>
-    );
-  };
+      </Grid>
+      <Divider />
+    </Box>
+  );
 
   return (
-    <React.Fragment>
+    <Root>
       {header()}
       <Divider />
       {processing && <LinearProgress />}
@@ -648,7 +697,9 @@ const TestDetailsModal: React.FunctionComponent<{
           container
           justifyContent="center"
           alignItems="stretch"
-          style={{ height: "100%" }}
+          style={{
+            height: "100%",
+          }}
         >
           {baselinePanel()}
           {isDiffShown
@@ -663,8 +714,8 @@ const TestDetailsModal: React.FunctionComponent<{
                 "Image",
                 testRun.branchName,
                 testRun.imageName,
-                imageStatus,
-                image,
+                currentImageStatus,
+                currentImage,
               )}
         </Grid>
       </Box>
@@ -673,6 +724,7 @@ const TestDetailsModal: React.FunctionComponent<{
           <Tooltip title={"Zoom In"}>
             <IconButton
               onClick={() => setStageScale(stageScale * stageScaleBy)}
+              size="large"
             >
               <ZoomIn />
             </IconButton>
@@ -680,17 +732,18 @@ const TestDetailsModal: React.FunctionComponent<{
           <Tooltip title={"Zoom Out"}>
             <IconButton
               onClick={() => setStageScale(stageScale / stageScaleBy)}
+              size="large"
             >
               <ZoomOut />
             </IconButton>
           </Tooltip>
           <Tooltip title={"Original size"}>
-            <IconButton onClick={setOriginalSize}>
+            <IconButton onClick={setOriginalSize} size="large">
               <Fullscreen />
             </IconButton>
           </Tooltip>
           <Tooltip title={"Fit into screen"}>
-            <IconButton onClick={fitImagesToStage}>
+            <IconButton onClick={fitImagesToStage} size="large">
               <FullscreenExit />
             </IconButton>
           </Tooltip>
@@ -704,8 +757,11 @@ const TestDetailsModal: React.FunctionComponent<{
           <Tooltip title={"Hotkey: ArrowLeft"}>
             <IconButton
               color="secondary"
-              style={{ visibility: currentRunIndex > 0 ? "visible" : "hidden" }}
+              style={{
+                visibility: currentRunIndex > 0 ? "visible" : "hidden",
+              }}
               onClick={handlePrevious}
+              size="large"
             >
               <NavigateBefore />
             </IconButton>
@@ -723,11 +779,12 @@ const TestDetailsModal: React.FunctionComponent<{
               color="secondary"
               style={{
                 visibility:
-                  currentRunIndex + 1 < totalTestRunCount
+                  totalTestRunCount > currentRunIndex + 1
                     ? "visible"
                     : "hidden",
               }}
               onClick={handleNext}
+              size="large"
             >
               <NavigateNext />
             </IconButton>
@@ -735,7 +792,6 @@ const TestDetailsModal: React.FunctionComponent<{
         </Grid>
         <Grid
           item
-          container
           xs={4}
           className={classes.testRunActions}
           justifyContent="flex-end"
@@ -772,7 +828,7 @@ const TestDetailsModal: React.FunctionComponent<{
           applyIgnoreArea();
         }}
       />
-    </React.Fragment>
+    </Root>
   );
 };
 
