@@ -1,43 +1,50 @@
 import React from "react";
+import { styled } from "@mui/material/styles";
 import { TestVariation } from "../types";
 import {
   Card,
   Grid,
   CardMedia,
   CardContent,
-  makeStyles,
   CardActions,
   Button,
   IconButton,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 import { staticService } from "../services";
 import { Link } from "react-router-dom";
 import { routes } from "../constants";
 import { TestVariationDetails } from "./TestVariationDetails";
-import { Delete } from "@material-ui/icons";
+import { Delete } from "@mui/icons-material";
 import { BaseModal } from "./BaseModal";
+
+const PREFIX = "TestVariationList";
+
+const classes = {
+  card: `${PREFIX}-card`,
+  media: `${PREFIX}-media`,
+};
+
+// TODO jss-to-styled codemod: The Fragment root was replaced by div. Change the tag if needed.
+const Root = styled("div")({
+  [`& .${classes.card}`]: {
+    maxWidth: 345,
+  },
+  [`& .${classes.media}`]: {
+    height: 140,
+    objectFit: "contain",
+  },
+});
 
 interface IProps {
   items: TestVariation[];
   onDeleteClick: (id: string) => void;
 }
 
-const useStyles = makeStyles({
-  card: {
-    maxWidth: 345,
-  },
-  media: {
-    height: 140,
-    objectFit: "contain",
-  },
-});
-
 const TestVariationList: React.FunctionComponent<IProps> = ({
   items,
   onDeleteClick,
 }) => {
-  const classes = useStyles();
   const [selectedItem, setSelectedItem] = React.useState<TestVariation | null>(
     null,
   );
@@ -47,7 +54,7 @@ const TestVariationList: React.FunctionComponent<IProps> = ({
   };
 
   return (
-    <React.Fragment>
+    <Root>
       <Grid container>
         {items.length === 0 && (
           <Typography variant="h5">No variations</Typography>
@@ -72,11 +79,7 @@ const TestVariationList: React.FunctionComponent<IProps> = ({
                 >
                   History
                 </Button>
-                <IconButton
-                  onClick={(event: React.MouseEvent<HTMLElement>) =>
-                    setSelectedItem(t)
-                  }
-                >
+                <IconButton onClick={() => setSelectedItem(t)} size="large">
                   <Delete />
                 </IconButton>
               </CardActions>
@@ -100,7 +103,7 @@ const TestVariationList: React.FunctionComponent<IProps> = ({
           }}
         />
       )}
-    </React.Fragment>
+    </Root>
   );
 };
 
