@@ -1,13 +1,11 @@
-import { Tooltip as MuiTooltip, type TooltipProps } from "@material-ui/core";
+import { Tooltip as MuiTooltip, TooltipProps } from "@mui/material";
 import React from "react";
 
 export const Tooltip: React.FunctionComponent<TooltipProps> = ({
   title,
   children,
-}) => {
-  return (
-    <MuiTooltip title={title} placement="bottom" arrow>
-      {children}
-    </MuiTooltip>
-  );
-};
+}) => (
+  <MuiTooltip title={title} placement="bottom" arrow>
+    {children}
+  </MuiTooltip>
+);
