@@ -1,33 +1,39 @@
 import React from "react";
-import { IconButton } from "@material-ui/core";
-import { Delete } from "@material-ui/icons";
+import { IconButton } from "@mui/material";
+import { Delete } from "@mui/icons-material";
 import {
-  type GridRowId,
-  useGridSlotComponentProps,
-} from "@material-ui/data-grid";
+  GridRowId,
+  useGridApiContext,
+  gridRowSelectionStateSelector,
+} from "@mui/x-data-grid";
 import { usersService } from "../../services";
 import { useSnackbar } from "notistack";
 import { useUserDispatch, useUserState } from "../../contexts";
 import { Tooltip } from "../Tooltip";
 
 export const ActionButtons: React.FunctionComponent = () => {
-  const { state } = useGridSlotComponentProps();
+  const apiRef = useGridApiContext();
+  const { state } = apiRef.current;
+  const selection = gridRowSelectionStateSelector(state);
+
   const { enqueueSnackbar } = useSnackbar();
   const userDispatch = useUserDispatch();
   const { userList, user } = useUserState();
 
   const ids: GridRowId[] = React.useMemo(
-    () => Object.values(state.selection),
-    [state.selection],
+    () => Object.values(selection),
+    [selection],
   );
+
   return (
-    <>
+    <React.Fragment>
       <Tooltip title="Delete selected rows." aria-label="delete">
         <span>
           <IconButton
             disabled={!ids.length}
             onClick={() => {
               const currentUserId = user?.id;
+
               if (currentUserId && ids.includes(currentUserId)) {
                 enqueueSnackbar(`You cannot delete yourself.`, {
                   variant: "error",
@@ -53,11 +59,12 @@ export const ActionButtons: React.FunctionComponent = () => {
                   );
               }
             }}
+            size="large"
           >
             <Delete />
           </IconButton>
         </span>
       </Tooltip>
-    </>
+    </React.Fragment>
   );
 };
