@@ -5,18 +5,24 @@ import { User } from "../types";
 export function authHeader() {
   // return authorization header with jwt token
   const userString = localStorage.getItem("user");
+
   if (userString) {
     const user: User = JSON.parse(userString);
-    if (user && user.token) {
-      return { Authorization: "Bearer " + user.token };
+
+    if (user?.token) {
+      return {
+        Authorization: "Bearer " + user.token,
+      };
     }
   }
+
   return [];
 }
 
 export function handleResponse(response: Response) {
   return response.text().then((text: string) => {
     const data = text && JSON.parse(text);
+
     if (!response.ok) {
       if (response.status === 401) {
         // auto logout if 401 response returned from api
@@ -27,7 +33,8 @@ export function handleResponse(response: Response) {
         }
       }
 
-      const error = (data && data.message) || response.statusText;
+      const error = data?.message || response.statusText;
+
       return Promise.reject(error);
     }
 
