@@ -1,4 +1,4 @@
-import { type Step } from "react-joyride";
+import { Step } from "react-joyride";
 
 export const LOCATOR_LOGIN_FORM = "loginform-1";
 export const LOCATOR_BUILD_DETAILS = "build-details";
@@ -37,43 +37,43 @@ export const PROJECT_LIST_PAGE_STEPS: Step[] = [
 
 export const PROJECT_PAGE_STEPS: Step[] = [
   {
-    target: "#" + LOCATOR_PROJECT_PAGE_SELECT_PROJECT,
+    target: `#${LOCATOR_PROJECT_PAGE_SELECT_PROJECT}`,
     content: "Select the project for which you want to view details.",
   },
   {
-    target: "#" + LOCATOR_PROJECT_PAGE_BUILD_LIST,
+    target: `#${LOCATOR_PROJECT_PAGE_BUILD_LIST}`,
     content: "List of Builds",
   },
   {
-    target: "#" + LOCATOR_PROJECT_PAGE_BUILD_LIST,
+    target: `#${LOCATOR_PROJECT_PAGE_BUILD_LIST}`,
     content:
       "If you see 'No Builds', please run your image comparison from any client.",
   },
   {
-    target: "#" + LOCATOR_PROJECT_PAGE_BUILD_DETAILS,
-    content: "Breif details for selected build.",
+    target: `#${LOCATOR_PROJECT_PAGE_BUILD_DETAILS}`,
+    content: "Brief details for selected build.",
   },
   {
-    target: "#" + LOCATOR_PROJECT_PAGE_TEST_RUN_LIST,
+    target: `#${LOCATOR_PROJECT_PAGE_TEST_RUN_LIST}`,
     content: "TestRuns for selected build.",
   },
 ];
 
 export const TEST_VARIATION_LIST_PAGE = [
   {
-    target: "#" + LOCATOR_TEST_VARIATION_LIST_PAGE_SELECT_PROJECT,
+    target: `#${LOCATOR_TEST_VARIATION_LIST_PAGE_SELECT_PROJECT}`,
     title:
       "Shows all the historical record of baselines by Name + Branch + OS + Browser + Viewport + Device",
     content: "Select the project you want to act on.",
   },
   {
-    target: "#" + LOCATOR_TEST_VARIATION_SELECT_BRANCH,
+    target: `#${LOCATOR_TEST_VARIATION_SELECT_BRANCH}`,
     title: "Merge from one branch to another",
     content:
       "Select the branch from/to which you want to merge the variations.",
   },
   {
-    target: "#" + LOCATOR_TEST_VARIATION_LIST_PAGE_SELECT_RESET_FILTER,
+    target: `#${LOCATOR_TEST_VARIATION_LIST_PAGE_SELECT_RESET_FILTER}`,
     content: "Only filtered items are displayed/merged to the target branch.",
   },
 ];
