@@ -1,19 +1,18 @@
 import React from "react";
-import { type Step } from "react-joyride";
+import { Step } from "react-joyride";
 
 interface ISetStepAction {
   type: "setSteps";
-  payload: Array<Step>;
+  payload: Step[];
 }
 
-type IAction = ISetStepAction;
-
-type Dispatch = (action: IAction) => void;
+type Dispatch = (action: ISetStepAction) => void;
 type State = {
-  helpSteps: Array<Step>;
+  helpSteps: Step[];
 };
-
-type HelpProviderProps = { children: React.ReactNode };
+type HelpProviderProps = {
+  children: React.ReactNode;
+};
 
 const StateContext = React.createContext<State | undefined>(undefined);
 const DispatchContext = React.createContext<Dispatch | undefined>(undefined);
@@ -22,13 +21,14 @@ const initialState: State = {
   helpSteps: [],
 };
 
-function reducer(state: State, action: IAction): State {
+function reducer(state: State, action: ISetStepAction): State {
   switch (action.type) {
     case "setSteps":
       return {
         ...state,
         helpSteps: action.payload,
       };
+
     default:
       return state;
   }
@@ -48,22 +48,29 @@ function HelpProvider({ children }: HelpProviderProps) {
 
 function useHelpState() {
   const context = React.useContext(StateContext);
+
   if (context === undefined) {
-    throw new Error("must be used within a HelpContext");
+    throw Error("must be used within a HelpContext");
   }
+
   return context;
 }
 
 function useHelpDispatch() {
   const context = React.useContext(DispatchContext);
+
   if (context === undefined) {
-    throw new Error("must be used within a HelpContext");
+    throw Error("must be used within a HelpContext");
   }
+
   return context;
 }
 
-function setHelpSteps(dispatch: Dispatch, data: Array<Step>) {
-  dispatch({ type: "setSteps", payload: data });
+function setHelpSteps(dispatch: Dispatch, data: Step[]) {
+  dispatch({
+    type: "setSteps",
+    payload: data,
+  });
 }
 
 export { HelpProvider, useHelpDispatch, useHelpState, setHelpSteps };
