@@ -10,7 +10,7 @@ interface IRequestAction {
 
 interface IGetAction {
   type: "get";
-  payload: Array<TestRun>;
+  payload: TestRun[];
 }
 
 interface ISelectAction {
@@ -20,17 +20,17 @@ interface ISelectAction {
 
 interface IDeleteAction {
   type: "delete";
-  payload: Array<string>;
+  payload: string[];
 }
 
 interface IAddAction {
   type: "add";
-  payload: Array<TestRun>;
+  payload: TestRun[];
 }
 
 interface IUpdateAction {
   type: "update";
-  payload: Array<TestRun>;
+  payload: TestRun[];
 }
 
 interface IApproveAction {
@@ -45,12 +45,12 @@ interface IRejectAction {
 
 interface IFilterAction {
   type: "filter";
-  payload?: Array<string | number>;
+  payload?: string[] | number[];
 }
 
 interface ISortAction {
   type: "sort";
-  payload?: Array<string | number>;
+  payload?: string[] | number[];
 }
 
 interface ITouchedAction {
@@ -70,18 +70,18 @@ type IAction =
   | ISortAction
   | ITouchedAction
   | ISelectAction;
-
 type Dispatch = (action: IAction) => void;
 type State = {
   selectedTestRun?: TestRun;
-  sortedTestRunIds?: Array<string | number>;
-  filteredTestRunIds?: Array<string | number>;
-  testRuns: Array<TestRun>;
+  sortedTestRunIds?: string[] | number[];
+  filteredTestRunIds?: string[] | number[];
+  testRuns: TestRun[];
   touched: boolean;
   loading: boolean;
 };
-
-type TestRunProviderProps = { children: React.ReactNode };
+type TestRunProviderProps = {
+  children: React.ReactNode;
+};
 
 const TestRunStateContext = React.createContext<State | undefined>(undefined);
 const TestRunDispatchContext = React.createContext<Dispatch | undefined>(
@@ -102,33 +102,39 @@ function testRunReducer(state: State, action: IAction): State {
         touched: false,
         selectedTestRun: action.payload,
       };
+
     case "filter":
       return {
         ...state,
         filteredTestRunIds: action.payload,
       };
+
     case "sort":
       return {
         ...state,
         sortedTestRunIds: action.payload,
       };
+
     case "request":
       return {
         ...state,
         testRuns: [],
         loading: true,
       };
+
     case "get":
       return {
         ...state,
         testRuns: action.payload,
         loading: false,
       };
+
     case "delete":
       return {
         ...state,
         testRuns: state.testRuns.filter((p) => !action.payload.includes(p.id)),
       };
+
     case "add":
       return {
         ...state,
@@ -140,25 +146,30 @@ function testRunReducer(state: State, action: IAction): State {
           ),
         ],
       };
+
     case "update":
       return {
         ...state,
         testRuns: state.testRuns.map((t) => {
           const item = action.payload.find((i) => i.id === t.id);
+
           if (item) {
             return item;
           }
+
           return t;
         }),
         selectedTestRun:
           action.payload.find((i) => i.id === state.selectedTestRun?.id) ??
           state.selectedTestRun,
       };
+
     case "touched":
       return {
         ...state,
         touched: action.payload,
       };
+
     default:
       return state;
   }
@@ -170,11 +181,17 @@ function TestRunProvider({ children }: TestRunProviderProps) {
 
   React.useEffect(() => {
     const { testId } = getQueryParams(location.search);
+
     if (!testId) {
-      dispatch({ type: "select" });
+      dispatch({
+        type: "select",
+      });
     } else {
       testRunService.getDetails(testId).then((payload) => {
-        dispatch({ type: "select", payload });
+        dispatch({
+          type: "select",
+          payload,
+        });
       });
     }
   }, [location.search]);
@@ -190,30 +207,43 @@ function TestRunProvider({ children }: TestRunProviderProps) {
 
 function useTestRunState() {
   const context = React.useContext(TestRunStateContext);
+
   if (context === undefined) {
-    throw new Error("must be used within a TestRunProvider");
+    throw Error("must be used within a TestRunProvider");
   }
+
   return context;
 }
 
 function useTestRunDispatch() {
   const context = React.useContext(TestRunDispatchContext);
+
   if (context === undefined) {
-    throw new Error("must be used within a TestRunProvider");
+    throw Error("must be used within a TestRunProvider");
   }
+
   return context;
 }
 
-async function deleteTestRun(dispatch: Dispatch, ids: Array<string>) {
-  dispatch({ type: "delete", payload: ids });
+function deleteTestRun(dispatch: Dispatch, ids: string[]) {
+  dispatch({
+    type: "delete",
+    payload: ids,
+  });
 }
 
-async function addTestRun(dispatch: Dispatch, testRuns: Array<TestRun>) {
-  dispatch({ type: "add", payload: testRuns });
+function addTestRun(dispatch: Dispatch, testRuns: TestRun[]) {
+  dispatch({
+    type: "add",
+    payload: testRuns,
+  });
 }
 
-async function updateTestRun(dispatch: Dispatch, testRuns: Array<TestRun>) {
-  dispatch({ type: "update", payload: testRuns });
+function updateTestRun(dispatch: Dispatch, testRuns: TestRun[]) {
+  dispatch({
+    type: "update",
+    payload: testRuns,
+  });
 }
 
 export {
